Tests/LibWeb/Text/expected/navigation/traverse-aborts-pending-intercepted-push.txt

navigate push destination=?pushed
push handler started
navigateerror at ?pushed
navigate traverse destination=?start
traverse handler
final location=?start index=0
